테스트용 업데이트

+++ src/main/java/com/ddiring/backend_asset/dto/UpdateAssetRequestDto.java
@@ -15,5 +15,5 @@ public class UpdateAssetRequestDto {
     private String buyAddress;
     private Long buyTokenAmount;
     private String sellAddress;
-    private Integer sellPrice; // 판매자에게 입금될 최종 금액
+    private Long sellPrice; // 판매자에게 입금될 최종 금액
 }
